requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
pgvector

openai

mlflow
apache-airflow

numpy
pandas
